I put in two important checks to ignore reservations currently in the system that violate our gap rules, as well as for cases like in Campsite 2 in our test file where the existing reservation's date range would actually fit inside our target reservation.
The CampsiteAvailability object is the 'manager' object that uses
Campsite and Reservation objects in order to calculate our available
campsites from the JSON input.  

In order to make date math easier to implement and maintain, I used Java
8 date classes.  Joda Time was considered, but the Java 8 classes were
sufficient to complete the task.  

Maven is used as the build system for both the main program and the test
suite.